really fixing signed unsigned issue in ftp-server (UINT8_MAX is often a signed integer constant)
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@104438 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r3/ftp-server.cpp
@@ -2269,7 +2269,7 @@ static int rtFtpServerCmdArgsParse(const char *pszCmdParms, uint8_t *pcArgs, cha
     int rc = RTGetOptArgvFromString(ppapszArgs, &cArgs, pszCmdParms, RTGETOPTARGV_CNV_QUOTE_MS_CRT, " " /* Separators */);
     if (RT_SUCCESS(rc))
     {
-        if ((unsigned int)cArgs <= UINT8_MAX)
+        if ((unsigned int)cArgs <= (unsigned int)UINT8_MAX)
         {
             *pcArgs = (uint8_t)cArgs;
         }
